+++ test/mitmproxy/contentviews/test_auto.py
@@ -47,7 +47,7 @@ def test_view_auto():
     assert f[0] == "Unknown Image"
 
     f = v(b"\xFF" * 30)
-    assert f[0] == "Hex"
+    assert f[0] == "Hexdump"
 
     f = v(
         b"",